% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_palette.R
\name{family_palette}
\alias{family_palette}
\title{Generate a color palette for nodes family}
\usage{
family_palette(family)
}
\arguments{
\item{family}{(list) : can be either the family column from the
legend table, or just a list of the families.
 In all cases, the parameter will be converted as a factor
 and sorted (alphabetically or numerically)
 Only its unique values are necessary.}
}
\value{
A list of key and values.
 - keys are the family names
 - values are the color
}
\description{
From a list of food families, create a color for each family.
}
\details{
Very useful when comparing graphs with the same families.
It can be used by itself, but this function was created to be the
\code{family_palette} argument when calling
\code{display_graph_from_links_nodes()} The colors will be
automatically added to the graph (nodes and legend)
}
\examples{
family_palette(c("Fruits", "Vegetables", "Meats"))
}
