% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mic_adj_matrix.R
\name{mic_adj_matrix}
\alias{mic_adj_matrix}
\title{Computes a MIC adjacency matrix}
\usage{
mic_adj_matrix(obs_data)
}
\arguments{
\item{obs_data}{(data.frame or matrix) : a dataset which rows are
observations and columns the variables.}
}
\value{
the adjacency matrix of MIC values for each pairwise association.
}
\description{
For a given dataset, computes the adjacency matrix
of maximal information coefficient (MIC) of each
pairwise association.
NOTE : another approach could have been to give the whole
 data frame to the \code{minerva} package func \code{cstats()},
 but it seemed slower in my tests.
}
\examples{
mic_adj_matrix(iris[,-5])
}
\references{
Reshef et al. (2011) <doi:10.1126/science.1205438>
}
