% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_graph.R
\name{save_graph}
\alias{save_graph}
\title{Save graph}
\usage{
save_graph(graph, filename = NULL, width = 25, height = 20,
  dpi = 300, ...)
}
\arguments{
\item{graph}{: the graph}

\item{filename}{(optional) : the name of the file. Default is the igraph unique ID in a .png format}

\item{width}{(optional) : width of the image in cm. Default is 25 cm}

\item{height}{(optional) : height of the image in cm. Default is 20 cm}

\item{dpi}{(optional) : the resolution of the image in dpi. Default is 300}

\item{...}{: other parameters to pass to the \code{ggsave} ggplot2 function}
}
\description{
Save the graph generated from \code{graph_from_matrix()} or \code{graph_from_links_nodes()}.
}
\examples{
adj_matrix <- cor(iris[,-5])
legend <- data.frame(name = colnames(iris[,-5]),
                     title = colnames(iris[,-5]))
graph_iris <- graph_from_matrix(adj_matrix, legend, main_title = "Iris graph")
save_graph(graph_iris)
}
